Moved GLTFLoader to auto-mock.
