+++ src/loadingmanager/__test__/LoadingManager.test.ts
@@ -1,25 +1,9 @@
 import { DIVELoadingManager } from '../LoadingManager';
 
-const mock_setDRACOLoader = jest.fn();
 const mock_setDecoderPath = jest.fn();
 
-jest.mock('three/examples/jsm/Addons.js', () => {
+jest.mock('three/examples/jsm/loaders/DRACOLoader', () => {
     return {
-        GLTFLoader: jest.fn(function () {
-            this.loadAsync = (
-                uri: string,
-                progEvent: (p: ProgressEvent<EventTarget>) => void,
-            ) =>
-                new Promise<void>((resolve) => {
-                    progEvent({
-                        loaded: 0,
-                        total: 1,
-                    } as ProgressEvent<EventTarget>);
-                    resolve();
-                });
-            this.setDRACOLoader = mock_setDRACOLoader;
-            return this;
-        }),
         DRACOLoader: jest.fn(() => {
             return {
                 setDecoderPath: mock_setDecoderPath,
@@ -32,8 +16,6 @@ describe('dive/loadingmanager/DIVELoadingManager', () => {
     it('should instantiate', () => {
         const testLight = new DIVELoadingManager();
         expect(testLight).toBeDefined();
-        expect(mock_setDecoderPath).toHaveBeenCalledTimes(1);
-        expect(mock_setDRACOLoader).toHaveBeenCalledTimes(1);
     });
 
     it('should return GLTF promise', () => {
