+++ src/model/__test__/Model.test.ts
@@ -2,7 +2,7 @@ import { RaycasterIntersectObjectMock } from '../../../__mocks__/three';
 
 import { DIVEModel } from '../Model';
 import { DIVECommunication } from '../../com/Communication';
-import { GLTF } from 'three/examples/jsm/Addons';
+import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
 import { DIVEScene } from '../../scene/Scene';
 import {
     Vector3,
